RESTful API endpoints for contact-messages

+++ pages/api/contact-messages/[id].js
@@ -2,11 +2,23 @@ import pg from "@/utils/pg";
 
 export default async function handler(req, res) {
     const { id } = req.query;
-    console.log(id);
+    
+    if (req.method === "DELETE") {
+        await pg.execute(`
+            delete from contact_messages where id = ${id}
+        `);
+        res.json({ message: "success" });
+        return;
+    }
 
-    await pg.execute(`
-        delete from contact_messages where id = ${id}
-    `);
-
-    res.json({ message: "success" })
+    if (req.method === "GET") {
+        const messages = await pg.execute(`
+            select * from contact_messages where id = '${id}'
+        `);
+        
+        res.json(messages[0]);
+        return;
+    }
+    
+    res.status(400).json({ message: "This method is not implemented" });
 }
